Implement better caching for requests
This is not a replacement for the caching system that's already in
place, because they have different reasons. The "old" is for making sure
that even if the network is offline, proxtop will be able to display the
last responses. For example when you close the app, went on a train ride
and opened the app again, then that cache is going to kick in.

This new caching system is for reducing load on the site and making proxtop respond
faster (with the cost of not being 100% up to date). The cache currently
- at least for me - is really passive. That means things are cached
  longer than you might want them to be cached, but I can improve on
this in the future as this is just the first step into a nicer and
better usability experience. The goal would be to directly display
information and update it shortly after if there's something newer
available.

This system only caches information that would get display, not actions
that you do, so those are guarateed to run.