Make sure to wait for a free limit even after delay
We have to (possibly) wait again after initially delaying the request
when reaching the API Limit.
Reason being that more requests get piled up than the next limit is
allowing to get through. By waiting for the limit again we ensure that
if more requests than the limit would allow got queued up, don't allow
all of them to go through.

This is subtle but is still important to not run into limits as we're
trying to avoid those.