Re-use old watchlist request and not api
For some reason the new watchlist API does not properly convey if an
episode is online or not. It has a "state" field, but this is for
the status of the anime itself (Airing, Finished, etc.) which does
not help here. So for now, we'll just use the old way of querying
the watchlist and switch over once there's a good way to do it.