Use webpack for building
This is the first in a serious of commits to add webpack to the project.
 This is the minimal required set of changes to make it work, but
obviously a lot more will be required to have it fully functional again.

The reasoning for this move is mostly to allow for moving towards
typescript and using webpack would be an easy solution to achieving
this. But since webpack is a lot more flexible than just for adding
typescript support, it is possible to add some other goodies, if
_really_ necessary.

I am aware that this makes it harder to develop and debug, but in this
case I think that the advantages outweigh the drawbacks, especially in
the long run.

Again, this commit only makes the application run again, but a few
things are missing. Most notably, the publishing currently won't work
and the process has to be remade. The tests that I have, albeit those
are only a few, have to be updated. Some other changes that aren't
really necessary but will be done are the removal of bower and the
upgrade of many dependencies to more recent versions. Typescript will
follow after that.