Move backend to typescript
This is the first big step in a long way. I'm planning to move the whole
 codebase to use typescript, both backend and frontend. This will be
mostly done in two parts, as the frontend would include moving to
angular 2+ as well. So for now, it's only (most of) the backend as this
can be done incrementally, though this is the first bulk of it.

The reason behind this move is 1) to allow for migrating to angular 2+
and 2) for a better development experience. There are so many different
API responses, IPC objects and the like getting moved around, it's hard
to keep track of them and is frustrating to work with. Since I got
webpack to work, adding typescript isn't such a big deal anymore, even
once I'll introduce it to the client side as well.