Fix mp4upload stream extraction not working
Mp4Upload has made it a little harder to get to the source url of the
video. The code they run on their page uses `eval` to run some
obfusecated code which boils down to replacing certain parts of a
template in a certain order with the values from a `|` separated list of
 strings. Luckily we can just extract each parameter and run it against
our own implementation of that replace function to avoid having to
resort to using `eval`. The resulting string is not actually the direct
source, but javascript code that would eventually set the source of the
video.