Automatically start next episode after certain time
You can now automatically start the next episode after certain amount
of seconds (before the end of the current episode). Similar to how
Netflix or Youtube do it. This can also be turned out to fall back to
manual switching. This does work in combination with automatically
marking the next episode as a watchlist item.