Use API endpoint to load watchlist
This also fixes mixing up EID and ID; the documentation is a little
confusing here. EID is described as the "ID of the Entry", whereas
"entry" refers to the anime/manga entry, _not_ the reminder entry.
However the description for the ID field here is unambiguous as the ID
of the reminder entry, so this corrects them.