Fix not being able to mark anime finished
This is mostly due to how this special case works. In this case,
setting the episode in the reminder doesn't work because we need to
specifically set the state of the comment to finished. If that was
already the case, we have to just delete the reminder, since the
comment already has the correct state.

We do this by querying all "currently watching" anime entries and
check whether there's one that matches the anime id. If we find one, we
 use the ID of that comment and set the new state. If we didn't, we
assume it's already finished.