Use webpack to replace api key
To make the build simpler and possibly eliminate the makefile in the
future, this removes one manual step that I always did; replacing the
proxer api key. Instead, it will be provided as a constant to the code
and compiled in by webpack, which already compiles all of the code
anyway.