Say goodbye to the Makefile
Personally, I like the makefile, but in order to make this more
portable it had to go. The goal is to make it easier for CI to run as
well as being able to use the publishing options provided by electron
builder. Since I was already using electron builder, just used it only
for building the final .exe, .zip, etc., there isn't much more setup to
 do thankfully.

Electron-builder advises the use of electron-webpack but I couldn't get
 it to work as it was always missing modules or other compile errors.
Since the current webpack setup works fine, I kept it.